fix(executor): handle empty-string organize path

+++ craft_parts/executor/organize.py
@@ -187,6 +187,9 @@ def organize_files(  # noqa: PLR0912, PLR0915
                         real_dst_path = dst_path.resolve()
                     else:
                         real_dst_path = dst_path
+                    if real_dst_path == src_path:
+                        # Trying to organize a directory to the same place, skipping
+                        continue
                     file_utils.link_or_copy_tree(src, real_dst_path)
                     shutil.rmtree(src)
                     continue
